requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.2
requests==2.31.0
langchain==0.1.0
langgraph==0.0.20
langchain-google-genai==0.0.5
google-genai==0.3.0
